Change ghc options
The rts options fix a possible crash when benchmarking multithreaded
code. The other options are an attempt to make the C code be faster, but
it doesn't seem to help too much. Manually compiling this _exact_ same
code and manually timing it shows that the speed is equivalent to the
Accelerate single-threaded code, but for some reason, I can't get stack
to compile it that fast.